Introduce Sapi Class
Move some logic from both Config and Utils that fit better here.
This should be considered very low level and should handle operating with the code in a very error safe way.

Additionally, fixed some warnings from static analysis.

+++ Sources/Graphics/Gif/Image.php
@@ -179,9 +179,9 @@ class Image
 				$this->m_data = substr($this->m_data, $this->m_gih->m_nWidth);
 
 				$data =
-					substr($data, 0, $y * $this->m_gih->m_nWidth) .
+					substr($data, 0, (int) ($y * $this->m_gih->m_nWidth)) .
 					$lne .
-					substr($data, ($y + 1) * $this->m_gih->m_nWidth);
+					substr($data, (int) (($y + 1) * $this->m_gih->m_nWidth));
 			}
 		}
